Fix: UPS cut glitch
Fix the glitch that cut the end of the file if it's larger than the changed file patch was originally created with like NUPS does, because many patch creators are using it to create UPS patches, and they won't apply correctly without this fix. It also fixes mismatching results with NUPS in the same file size.